% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblegrid.R
\name{bubblegrid}
\alias{bubblegrid}
\title{Bubble grid plots (for single-cell data)}
\usage{
bubblegrid(x, y, color, num_bins = 25)
}
\arguments{
\item{x, y, color}{Numeric vectors. To be shown directly in the plot.}

\item{num_bins}{This function works by hexagonal binning. This is the number of bins on the horizontal axis.}
}
\description{
Bubble grid plots (for single-cell data)
}
