% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblegrid.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Bubblegrid test dataset}
\format{An object of class \code{data.frame} with 1891 rows and 4 columns.}
\usage{
test_data
}
\description{
A small dataset with 2D Monocle embedding and two genes.
}
\keyword{datasets}
