% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblegrid.R
\name{bubblegrid}
\alias{bubblegrid}
\title{Print and return bubble grid plots (for single-cell data)}
\usage{
bubblegrid(x, y, color, num_bins = 25)
}
\arguments{
\item{x, y, color}{Numeric vectors. To be shown directly in the plot.}

\item{num_bins}{This function works by hexagonal binning. This is the number of bins on the horizontal axis.}
}
\value{
Return value is a ggplot object.
}
\description{
Print and return bubble grid plots (for single-cell data)
}
\examples{
data(test_data)
bubblegrid::bubblegrid(x = test_data[[1]], y = test_data[[2]], color = test_data[[3]])

}
